# api.yaml
openapi: 3.0.3
info:
  title: API for project 3d-inventory
  description: >-
    REST API for **3d-inventory**. Project is a simple solution that allows you to build a spatial and database
    representation of all types of warehouses and server rooms.
  contact:
    name: 3d-inventory-mongo-api
    url: https://github.com/karol-preiskorn/3d-inventory-mongo-api/
  version: 0.54.0
servers:
  - url: https://127.0.0.1:8080
  - url: http://0.0.0.0:8080
  - url: https://0.0.0.0:8080
  - url: http://localhost:8080
  - url: https://localhost:8080
  - url: http://127.0.0.1:8080
  - url: http://172.17.0.2:8080
  - url: http://172.17.0.3:8080
  - url: http://172.20.0.3:8080
  - url: https://d-inventory-api-wzwe3odv7q-ew.a.run.app
  - url: https://3d-inventory-api.ultimasolution.pl
  - url: https://3d-inventory.ultimasolution.pl
...
paths:
...
  /devices:
    get:
      tags:
        - devices
      summary: Get all devices
      description: Retrieve a list of all devices.
      operationId: getAllDevices
      security:
        - ViewerAuth: []
      responses:
        '200':
          description: A list of devices
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (viewer role or higher required)
        '404':
          description: No devices found.
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        '500':
          description: Unexpected server error
        default:
          description: Unexpected error
      x-swagger-router-controller: Devices
    put:
      tags:
        - devices
      summary: Update an existing device
      description: Update the details of an existing device.
      operationId: updateDevice
      security:
        - UserAuth: []
      responses:
        '200':
          description: Device updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (user role or higher required)
        '404':
          description: Device update data is inconsistent.
        '406':
          description: Not Acceptable
        '429':
          description: Too Many Requests
        '500':
          description: Unexpected server error
        default:
          description: Unexpected error
      x-swagger-router-controller: Devices
    delete:
      tags:
        - devices
      summary: Delete all devices
      description: Remove all devices from the inventory.
      operationId: deleteAllDevices
      security:
        - AdminAuth: []
      responses:
        '200':
          description: All devices deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmation'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: No devices found to delete
      x-swagger-router-controller: Devices
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication. Include role and permissions for authorization.
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with admin role required
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with user role or higher required
    ViewerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with viewer role or higher required (read-only access)
  schemas:
...
    Attribute:
      description: >-
        Attribute is a key-value pair that can be associated with a device, model, or connection.
      x-42c-format: o:attribute
      required:
        - _id
        - attributeDictionaryId
        - value
      type: object
      properties:
        _id:
          example: 6566ed4ace63d42b25fbba19
          pattern: ^[0-9A-Fa-f]{24}$
          minLength: 24
          maxLength: 24
          x-42c-format: o:uuid
          type: string
        attributeDictionaryId:
          type: string
        connectionId:
          type: string
        deviceId:
          type: string
        modelId:
          type: string
        value:
          type: string
      example:
        _id: 653e10ecd6025afb66017c65
        attributeDictionaryId: 653e10ecd6025afb66017c65
        connectionId: 653e10ecd6025afb66017c70
        deviceId: 653e10ecd6025afb66017c71
        modelId: 653e10ecd6025afb66017c72
        value: 3W
    Attributes:
      type: array
      description: List of attributes
      maxItems: 1000
      example:
        - _id: 653e10ecd6025afb66017c65
          attributeDictionaryId: 653e10ecd6025afb66017c65
          connectionId: 653e10ecd6025afb66017c70
          deviceId: 653e10ecd6025afb66017c71
          modelId: 653e10ecd6025afb66017c72
          name: power
          value: 3W
        - _id: 653e10ecd6025afb66017c66
          attributeDictionaryId: 653e10ecd6025afb66017c66
          connectionId: 653e10ecd6025afb66017c73
          deviceId: 653e10ecd6025afb66017c74
          modelId: 653e10ecd6025afb66017c75
          name: color
          value: black
        - _id: 653e10ecd6025afb66017c67
          attributeDictionaryId: 653e10ecd6025afb66017c67
          connectionId: 653e10ecd6025afb66017c76
          deviceId: 653e10ecd6025afb66017c77
          modelId: 653e10ecd6025afb66017c78
          name: year
          value: '1999'
      items:
        $ref: '#/components/schemas/Attribute'
    Position:
      required:
        - x
        - 'y'
        - h
      type: object
      properties:
        x:
          type: number
          description: Width
        'y':
          type: number
          description: Depth
        h:
          type: number
          description: Height
      additionalProperties: false
      description: Position device on the floor
    Device:
      required:
        - _id
        - modelId
        - name
      type: object
      properties:
        _id:
          example: 6566ed4ace63d42b25fbba19
          pattern: ^[0-9A-Fa-f]{24}$
          minLength: 24
          maxLength: 24
          x-42c-format: o:uuid
        name:
          type: string
          description: Device name
        modelId:
          example: 6566ed4ace63d42b25fbba19
          type: string
          pattern: ^[0-9A-Fa-f]{24}$
          minLength: 24
          maxLength: 24
        position:
          $ref: '#/components/schemas/Position'
        attributes:
          $ref: '#/components/schemas/Attributes'
      description: Device it is basic structure 3d-inventory
      example:
        _id: 6566ed4ace63d42b25fbba19
        name: Device 2
        modelId: 6566ed4ace63d42b25fbba19
        position:
          x: 4
          'y': 3
          h: 0
        attributes:
          - _id: 653e10ecd6025afb66017c65
            key: power
            value: 3W
          - _id: 653e10ecd6025afb66017c66
...
    Confirmation:
      type: object
      properties:
        message:
          type: string
      additionalProperties: false
      description: Universal confirmation message from database
      example:
        message: Succeed
...